+++ development/Sandbox/Program.cs
@@ -6,13 +6,13 @@ using Sandbox.Providers;
 Console.Clear();
 
 Console.OutputEncoding = Encoding.UTF8;
-//await Sandbox.Providers.Anthropic.RunAsync();
+await Sandbox.Providers.AnthropicDemo.RunAsync();
 //await Sandbox.Providers.AmazonBedrock.RunAsync();
 //await Sandbox.Providers.OpenAI.RunAsync();
 //await Sandbox.Providers.GitHub.RunAsync();
 //await AzureOpenAI.RunAsync();
 //await Sandbox.Providers.Mistral.RunAsync();
-await Sandbox.Providers.Google.RunAsync();
+//await Sandbox.Providers.Google.RunAsync();
 //await Sandbox.Providers.XAI.RunAsync();
 //await Sandbox.Providers.OpenRouter.RunAsync();
 //await Sandbox.Providers.Cerebras.RunAsync();